GenAI_Projects/Proj-6_Pinecone_DB/requirements.txt
langchain
pypdf
openai
tiktoken
langchain_community
pinecone
python-dotenv
werkzeug